+++ Sources/SourceGraph/SourceGraphMutatorRunner.swift
@@ -33,6 +33,7 @@ public final class SourceGraphMutatorRunner {
         EnumCaseReferenceBuilder.self,
         DefaultConstructorReferenceBuilder.self,
         StructImplicitInitializerReferenceBuilder.self,
+        InheritedImplicitInitializerReferenceBuilder.self,
 
         DynamicMemberRetainer.self,
         UnusedParameterRetainer.self,
@@ -42,6 +43,7 @@ public final class SourceGraphMutatorRunner {
         XCTestRetainer.self,
         SwiftTestingRetainer.self,
         SwiftUIRetainer.self,
+        AppIntentsRetainer.self,
         StringInterpolationAppendInterpolationRetainer.self,
         PropertyWrapperRetainer.self,
         ResultBuilderRetainer.self,